% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-case.R
\name{change_output_ntstep}
\alias{change_output_ntstep}
\title{This function change the output time steps}
\usage{
change_output_ntstep(case.list = NULL, sobek.project = NULL,
  n.tstep = NULL)
}
\arguments{
\item{case.list}{List of cases to change}

\item{sobek.project}{Path to Sobek project}

\item{n.tstep}{An integer - Number of time step to export result}
}
\description{
This function change the output time steps
}
