% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobek-profile.R
\name{change_rl_rs}
\alias{change_rl_rs}
\title{Change reference level of a cross section}
\usage{
change_rl_rs(crsn.id, pf.df, rl.new = 0, rs.new = rl.new + 10)
}
\arguments{
\item{crsn.id}{ID of cross section}

\item{rl.new}{New value of bed level left}

\item{rs.new}{New value of surface level right, default = rl.new + 10}
}
\description{
Change reference level of a cross section
}
