% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-case.R
\name{create_case}
\alias{create_case}
\title{Create new Sobek Case from existing Case}
\usage{
create_case(old.name = "", new.name = "",
  sobek.project = "d:/so21302/rhein29A.lit", new.begin = NULL,
  new.end = NULL, new.desc = NULL)
}
\arguments{
\item{old.name}{Name of the old case, case-sensitive}

\item{new.name}{Name of the new case}

\item{sobek.project}{Sobek Project folder}

\item{new.begin}{New simulation starting time (in format dd.mm.yyyy hh:mm:ss)}

\item{new.end}{New simulation starting time (in format dd.mm.yyyy hh:mm:ss)}

\item{new.desc}{Character vector, for new case description, one element for each line}
}
\description{
Create new Sobek Case from existing Case
}
