% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-dat.R
\name{create_dat}
\alias{create_dat}
\title{Create Sobek Input .DAT from}
\usage{
create_dat(id.file = "", data.file = "", bnd.header = NULL,
  lat.header = NULL, factor = 1, bnd.out = NULL,
  lat.out = sub("boundary.dat", "lateral.dat", bnd.out, ignore.case =
  TRUE, fixed = TRUE), col.sep = "\\t", dec.char = ",",
  output.case = NULL)
}
\arguments{
\item{id.file}{Path to ID file}

\item{data.file}{Path to Data file}

\item{bnd.header}{Path to header file for boundary}

\item{lat.header}{Path to header file for Lateral}

\item{factor}{Vergrößerung Faktor}

\item{bnd.out}{Output path for Boundary.dat (full file path)}

\item{lat.out}{Output path for Lateral.dat (full file path)}

\item{col.sep}{Column seperator of ID & Input Data files. Default TAB.}

\item{dec.char}{Decimal character of ID & Input Data files. Default ","}

\item{output.case}{Write output (boundary.dat, lateral.dat) direct to the case.
Output.case must be a character vector combine (case.name, sobek.project)
like c('case name', 'd:/so21302/main.lit')}
}
\description{
Create Sobek Input .DAT from
}
\details{
Example for id.file
\tabular{llllrr}{
Ahr           \tab ahr_mess_fv_q         \tab FLBR id Ahr sc 0 lt 0 dc lt 1 0 0 PDIN 0 0  pdin              \tab Lateral  \tab 1 \tab 100\cr
Mai_23        \tab bad_vibel_mess_fv_q   \tab FLBO id Mai_23 st 0 ty 1 q_ dt 1 0 0 PDIN 0 0  pdin           \tab Boundary \tab 1 \tab 100\cr
L_10014432    \tab cochem_mess_q         \tab FLBO id Mosel_RE_N_P_P_0 st 0 ty 1 q_ dt 1 0 0 PDIN 0 0  pdin \tab Boundary \tab 1 \tab 100\cr
Modau         \tab modau_mess_fv_q       \tab FLBR id Modau sc 0 lt 0 dc lt 1 0 0 PDIN 0 0  pdin            \tab Lateral  \tab 1 \tab 100\cr
Mai_Lat_Schwb \tab schwarzbach_mess_fv_q \tab FLBR id Mai_Lat_Schwb sc 0 lt 0 dc lt 1 0 0 PDIN 0 0  pdin    \tab Lateral  \tab 1 \tab 100
}
}
