% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.his_parameter}
\alias{.his_parameter}
\title{Get parameter table from .HIS & .HIA files}
\usage{
.his_parameter(his.file = "")
}
\arguments{
\item{his.file}{Path to the .HIS file}
}
\value{
a data.table with two column: location & sobek.id
}
\description{
Get parameter table from .HIS & .HIA files
}
