% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{.id2loc}
\alias{.id2loc}
\title{Convert Sobek ID to location index}
\usage{
.id2loc(id, his.locs)
}
\arguments{
\item{his.locs}{Location table}

\item{sobek.id}{Node/Reach ID}
}
\description{
Convert Sobek ID to location index
}
