% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobek_export.R
\name{export_sobek}
\alias{export_sobek}
\title{Export list of cases from a sobek project to another location}
\usage{
export_sobek(case.list, sobek.project, dest)
}
\arguments{
\item{case.list}{Path to file with case list}

\item{sobek.project}{Path to current sobek project}

\item{dest}{Destination folder}
}
\description{
Export list of cases from a sobek project to another location
}
