% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_column_by_name}
\alias{get_column_by_name}
\title{Return data for one column in a data table with value stored in the column parameter}
\usage{
get_column_by_name(column = NULL, df = NULL, suffix = "",
  all.lower = F, short = 0L)
}
\arguments{
\item{column}{Name of the column}

\item{df}{a data.table or data.frame}

\item{suffix}{Adding suffix to the column name}

\item{all.lower}{Should column converted to lower case? Default = FALSE#}

\item{short}{Get only a given number of character, from left, of the column name}
}
\value{
behave like column subsetting of data.table/data.frame
}
\description{
Return data for one column in a data table with value stored in the column parameter
}
