% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change-dat.R
\name{get_data_from_id}
\alias{get_data_from_id}
\title{Get the Data table for an element with specific ID from a .DAT files}
\usage{
get_data_from_id(dat.file = NULL, s.id = NULL)
}
\arguments{
\item{dat.file}{Path to .DAT file}

\item{s.id}{ID of the element to get data}
}
\value{
a data.table
}
\description{
Get the Data table for an element with specific ID from a .DAT files
}
