% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobek-case-info.R
\name{get_file_path}
\alias{get_file_path}
\title{Get file path from sobek case name}
\usage{
get_file_path(case.name = NULL, sobek.project = NULL, type = NULL)
}
\arguments{
\item{case.name}{Name of the case}

\item{sobek.project}{Path to Sobek project folder}

\item{type}{Type of file to get path (lateral, boundary, reach, node, structure...)}
}
\description{
Get file path from sobek case name
}
