% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-his.R
\name{his_from_file}
\alias{his_from_file}
\title{Export data for nodes/reaches using IDs from a file}
\usage{
his_from_file(his.file, id.file, param = 1L, f.header = FALSE,
  f.sep = "\\t", f.comment = "#")
}
\arguments{
\item{his.file}{Path to the .HIS file}

\item{id.file}{Path to file contain list of IDs.
ID file should have one-two columns, first for the sobek ID and the second (optional) for names that will be column names in the output data.table.}

\item{param}{Index/Name of the Paramter to get the data, default = 1}

\item{f.header}{The id.file contain header?, boolean, default = FALSE}

\item{f.sep}{Seperator of the id.file, default = 'TAB'}

\item{f.comment}{Commment character of the file, default = '#'}
}
\value{
A data.table
}
\description{
Export data for nodes/reaches using IDs from a file
}
