% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-his.R
\name{his_from_file_1}
\alias{his_from_file_1}
\title{Export data for nodes/reaches using IDs from a file}
\usage{
his_from_file_1(his.file, id.file, f.header = FALSE, f.sep = "\\t",
  param = 1L)
}
\arguments{
\item{his.file}{Path to the .HIS file, string}

\item{id.file}{Path to file contain list of IDs \n
IDs list shoulde have the first column with IDs(col.name = "sobek.id")\n
Second column (col.name = "name") is for Names correspond to the IDs}

\item{f.header}{If the id.file contain header?, boolean, default = FALSE}

\item{f.sep}{Seperator of the id.file, default = '\t'}

\item{param}{Index of the Paramter to get the data, default = 1}
}
\value{
A data.frame
}
\description{
Export data for nodes/reaches using IDs from a file
}
