% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-his.R
\name{his_from_list}
\alias{his_from_list}
\title{Export data for nodes/reaches using IDs from a list}
\usage{
his_from_list(his.file = NULL, id.list = NULL, param = 1L)
}
\arguments{
\item{his.file}{Path to the .HIS file, string}

\item{id.list}{List of Sobek IDs}

\item{param}{Index or Name of the Paramter to get the data, default = 1.}
}
\value{
A data.table
}
\description{
Export data for nodes/reaches using IDs from a list
}
