% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-his.R
\name{his_from_list_2}
\alias{his_from_list_2}
\title{Export data for nodes/reaches using IDs from a list}
\usage{
his_from_list_2(his.file = NULL, id.list = NULL, param = 1L)
}
\arguments{
\item{his.file}{Path to the .HIS file, string}

\item{id.list}{List of IDs to search for data
IDs list shoulde have the first column with IDs(col.name = "sobek.id")\cr
Second column (col.name = "name") is for Names correspond to the IDs\cr}

\item{param}{Index of the Paramter to get the data, default = 1\cr
Name of the parameter can be given instead of index}
}
\value{
A data.frame
}
\description{
Export data for nodes/reaches using IDs from a list
}
