% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-his.R
\name{his_info}
\alias{his_info}
\title{Read basic information of the .HIS file}
\usage{
his_info(his.file = "")
}
\arguments{
\item{his.file}{Path to the .HIS file}
}
\value{
A list of basic information, includes:
\itemize{
\item Name of the Sobek Modell Case: case_name
\item Number of parameters: param_nr
\item Names of Parameters: param_names
\item Total locations (ID): total_loc
\item Total time steps: total_tstep
\item Start time ot the timeserie: his_t0
\item Time step of the timeserie: his_dt (sec.)
}
}
\description{
Read basic information of the .HIS file
}
