% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/his-single-case.R
\name{his_single_case}
\alias{his_single_case}
\title{Export data for nodes/reaches for one SOBEK Case}
\usage{
his_single_case(case.name = NULL, sobek.project = NULL, qID = NULL,
  wID = NULL, lID = NULL, sID = NULL, mID = NULL, f.sep = "\\t",
  f.header = FALSE, param = 1, verbose = FALSE)
}
\arguments{
\item{case.name}{Name of the SOBEK case to work with\cr}

\item{qID}{Path to file containing list of reach IDs for struct.his\cr}

\item{wID}{IDs for calcpnt.his\cr}

\item{lID}{IDs for lateral.his\cr}

\item{sID}{IDs for struct.his\cr}

\item{mID}{IDs for measstat.his\cr}

\item{f.sep}{Seperator of the id.file, default = '\\t'}

\item{f.header}{Logical. Has id.file header?}

\item{param}{Index/Name of the Paramter to get the data, default = 1}

\item{verbose}{Should message be displayed?}
}
\value{
A list of data.table
}
\description{
For all .HIS files subfolders.
}
