% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_locs}
\alias{plot_locs}
\title{Plot graphic for some locations for one or more cases}
\usage{
plot_locs(case.names = "", sobek.project = "", IDs = "",
  type = c("node", "reach", "mea"))
}
\description{
Plot graphic for some locations for one or more cases
}
