% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pegel.R
\name{plot_pegel}
\alias{plot_pegel}
\title{Ganglinien an jedem Pegel (3 Linie pro Darstellung)}
\usage{
plot_pegel(pos = "Trunstadt", indt, x.lab = "Zeit",
  y.lab = "Abfluss (m³/s)", y2.scale = 2,
  title = paste(str_extract(y.lab, "[^ ]*"), "Ganglinien am Pegel:",
  pos), facet = TRUE, facet_list = list(hwe = "free_x"),
  vgf.pattern = "Selten|Mittel", zustand1 = "Bezugszustand",
  zustand2 = "Planzustand", sep.delta = FALSE)
}
\arguments{
\item{pos}{Name of Pegel}

\item{indt}{Input data.table}

\item{x.lab}{X-Axis label, default 'Zeit'}

\item{y.lab}{Y-Axis label, default 'Abfluss (m³/s)'}

\item{y2.scale}{Scale between y_max/y2_max}

\item{title}{Title of the graphic, by default: automatic extract from y.lab and pos}

\item{facet}{Should grahic be factted. Default TRUE}

\item{facet_list}{List of facetting parameters (ex. list(hwe = 'free_x'))
This parameter is currently not fully implemented}

\item{vgf.pattern}{Regex pattern to search for VGF}

\item{zustand1}{Name of Zustand 1}

\item{zustand2}{Name of Zustand 2. The Delta will be Zustand2 - Zustand1}

\item{sep.delta}{Should Delta be drew on a second graphic underneath}
}
\value{
A ggplot2 graphic
}
\description{
Ganglinien an jedem Pegel (3 Linie pro Darstellung)
}
