% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scheitel_delta}
\alias{plot_scheitel_delta}
\title{Plot peak difference at one place}
\usage{
plot_scheitel_delta(indt = NULL, pos = "", peak.duration = 5L,
  x.lab = "Zeit", y.lab = "Discharge (m³/s)",
  title = paste("Ganglinie an der Lage:", pos), x.angle = 90L,
  x.hjust = 1L, size = 1L, y.min = NULL, delta = TRUE,
  h.line = NULL, y.interval = 500L, faced = FALSE)
}
\arguments{
\item{indt}{Input data}

\item{pos}{Names of columns to draw}

\item{peak.duration}{Halftime of the flooding period, default 5 days}

\item{x.lab}{Label for x axis, default 'Lage'}

\item{y.lab}{Label for y axis, default 'Discharge (m³/s)'}

\item{title}{Graphic title}

\item{x.angle}{Rotation angle of x label}

\item{x.hjust}{hjust of x label}

\item{y.min}{ymin}

\item{delta}{Annotate Delta value, if only for one case and two columns}

\item{h.line}{Numeric vector for adding HLINE}

\item{y.interval}{Tick interval for y axis}

\item{faced}{faced by case, default FALSE}
}
\value{
a ggplot2 graphic
}
\description{
Plot peak difference at one place
}
