% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_wirkung}
\alias{plot_wirkung}
\title{Quick plot for long profile or pos (column name)}
\usage{
plot_wirkung(indt = NULL, pos = "longprofile", x.lab = "Lage",
  y.lab = "Wasserstand (m+NHN)", title = "", x.angle = 90L,
  x.hjust = 1L, size = 1L, faced = FALSE, sort = TRUE)
}
\arguments{
\item{indt}{Input data.table or data.frame}

\item{pos}{"longprofile" or list of list of column to plot}

\item{x.lab}{x label}

\item{y.lab}{y label}

\item{title}{title}

\item{x.angle}{rotate angle of x.label}

\item{x.hjust}{hjust of x.label}

\item{size}{size of line/point}

\item{faced}{default FALSE. If TRUE, faced by case}

\item{sort}{to short longprofile by max value, not yet implemented}
}
\value{
a ggplot2 graphic object
}
\description{
Quick plot for long profile or pos (column name)
}
