% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polder_tmv.R
\name{polder_tmv}
\alias{polder_tmv}
\title{Search for a polder TMV}
\usage{
polder_tmv(volume = NULL, indt = NULL, n.days = 2,
  value.step = 0.1, tolerance = 0.01, clipboard = TRUE,
  print.plot = FALSE, ...)
}
\arguments{
\item{volume}{Volume of the Polder}

\item{indt}{The discharge data.table. Only the first two columns will be used. It should have the format 'ts|value'}

\item{n.days}{Maximum number of days under the flood peak should the topping value being searched}

\item{value.step}{The discharge will be searched from the minimum to maximum by this value.step}

\item{tolerance}{The acceptable tolerance of calculated Volume and given Volume}

\item{clipboard}{Should the output table write to clipboard, ready for pasting in Sobek? Default TRUE}

\item{print.plot}{Should a plot be showed? Default FALSE}

\item{...}{If indt is NULL these parameters will be passed to his_from_case function for reading the discharge}
}
\description{
Search for a polder TMV
}
