% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change-dat.R
\name{set_const_id}
\alias{set_const_id}
\title{Set Value of an element in .DAT file to constant, and remove the time serie table}
\usage{
set_const_id(dat.file = NULL, s.id = NULL, value = 0,
  output = NULL)
}
\arguments{
\item{dat.file}{Path to .DAT file}

\item{s.id}{ID of the element to get data}

\item{value}{The constant value to set}

\item{output}{Path to write output file. Use the same path as dat.file to overwrite}
}
\description{
Set Value of an element in .DAT file to constant, and remove the time serie table
}
