% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobek-profile.R
\name{set_weir_pfd}
\alias{set_weir_pfd}
\title{Set possible flow  direction. Use it to close or open a simple weir}
\usage{
set_weir_pfd(weir.id, f.struct, direction = 1L)
}
\arguments{
\item{weir.id}{ID of the weir}

\item{f.struct}{Path toe control.def file, get it by function get_file_path}

\item{direction}{Possible flow direction (0-both, 1-positive, 2-negative, 3-no flow)}
}
\description{
Set possible flow  direction. Use it to close or open a simple weir
}
