% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobek-case-info.R
\name{sobek_case_info}
\alias{sobek_case_info}
\title{Read .HIS information for a case}
\usage{
sobek_case_info(case.name = NULL, sobek.project = NULL,
  his.type = NULL, info = NULL)
}
\arguments{
\item{case.name}{Name of Sobek case}

\item{sobek.project}{Sobek Folder}

\item{his.type}{c("node", "reach", "lateral", "struct", "measstation")}

\item{info}{c("location", "general")}
}
\description{
Read .HIS information for a case
}
