% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobek_export.R
\name{sobek_export}
\alias{sobek_export}
\title{Export list of cases from a sobek project to another location}
\usage{
sobek_export(case.list, sobek.project, dest)
}
\arguments{
\item{case.list}{Path to file with case list}

\item{sobek.project}{Path to current sobek project}

\item{dest}{Destination folder}
}
\description{
Export list of cases from a sobek project to another location
}
