% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobek-simulation.R
\name{sobek_view}
\alias{sobek_view}
\title{Run Sobek Simulation for single case}
\usage{
sobek_view(case.name = NULL, sobek.project = NULL, sobek.path = NULL,
  overwrite = TRUE, clear.temp = TRUE,
  cmt.template = "c:/so21503/cmtwork")
}
\arguments{
\item{case.name}{Name of the Case}

\item{sobek.project}{Path to Sobek Project Folder}

\item{sobek.path}{Path to Sobek Program Folder (ex. d:/so21302)}

\item{overwrite}{Should simulation result overwrite back to Case folder?}

\item{clear.temp}{Should temp directory to be clear?}
}
\description{
Run Sobek Simulation for single case
}
