% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobek-profile.R
\name{switch_DRV}
\alias{switch_DRV}
\title{Change profile definitions for DRV}
\usage{
switch_DRV(case.name, sobek.project, drv.name = NULL, status = "mit",
  f.tbl)
}
\arguments{
\item{f.tbl}{Path to profile definition table}

\item{case.name:}{Name of case}

\item{sobek.project:}{Path to Sobek Project}

\item{drv.name:}{Name of DRV}
}
\value{
Result will be written directly to the profile.dat
}
\description{
Change profile definitions for DRV
}
