% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-dat.R
\name{transfer_fra}
\alias{transfer_fra}
\title{Change Rhein Modell Boundary 'Frankfurt Ost (66) by ts from Main Output
This function is to extract Main Model output at Frankfurt Osthafen
and paste it as input to Rhein Model (boundary node '66')}
\usage{
transfer_fra(main.case = "", rhein.case = "",
  main.prj = "d:/so21302/main2015.lit",
  rhein.prj = "d:/so21302/rhein29a.lit")
}
\arguments{
\item{main.case}{Name of case in Main Model}

\item{rhein.case}{Name of case in Rhein Model}

\item{main.prj}{Path to Main Model (default d:/so21302/main2015.lit)}

\item{rhein.prj}{Path to Rhein Model (default d:/so21302/rhein29a.lit)}
}
\value{
Modified boundary.dat for the rhein.case
}
\description{
Change Rhein Modell Boundary 'Frankfurt Ost (66) by ts from Main Output
This function is to extract Main Model output at Frankfurt Osthafen
and paste it as input to Rhein Model (boundary node '66')
}
