% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobek-profile.R
\name{change_rl_rs}
\alias{change_rl_rs}
\title{Change reference level of a cross section}
\usage{
change_rl_rs(id, case.name = NULL, sobek.project = NULL, rl.new = 0,
  rs.new = rl.new + 10, profile.dat = NULL)
}
\arguments{
\item{id}{ID of cross section}

\item{case.name}{Name of the case}

\item{sobek.project}{Path to sobek.project}

\item{rl.new}{New value of bed level left}

\item{rs.new}{New value of surface level right, default = rl.new + 10}

\item{profile.dat}{Path to the profile.dat, if case and project were not given}
}
\description{
Change reference level of a cross section
}
