% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change-dat.R
\name{change_tble}
\alias{change_tble}
\title{Set Value of an element in .DAT file to constant, and remove the time serie table}
\usage{
change_tble(dat.file = NULL, s.id = NULL, tble = NULL,
  output = NULL)
}
\arguments{
\item{dat.file}{Path to .DAT file}

\item{s.id}{ID of the element to get data}

\item{tble}{data.frame with in format "ts|value"}

\item{output}{Path to write output file. Use the same path as dat.file to overwrite}
}
\description{
Set Value of an element in .DAT file to constant, and remove the time serie table
}
