% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-dat.R
\name{create_dat}
\alias{create_dat}
\title{Create Sobek Input .DAT from}
\usage{
create_dat(id.file = "", data.file = "", bnd.header = NULL,
  lat.header = NULL, factor = 1L, bnd.out = NULL, lat.out = NULL,
  col.sep = "\\t", dec.char = ",")
}
\arguments{
\item{id.file}{Path to ID file}

\item{data.file}{Path to Data file}

\item{bnd.header}{Path to header file for boundary}

\item{lat.header}{Path to header file for Lateral}

\item{factor}{Vergrößerung Faktor}

\item{bnd.out}{Output path for Boundary.dat (full file path)}

\item{lat.out}{Output path for Lateral.dat (full file path)}

\item{col.sep}{Column seperator of ID & Input Data files. Default TAB.}

\item{dec.char}{Decimal character of ID & Input Data files. Default ","}
}
\description{
Create Sobek Input .DAT from
}
