% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_and_import.R
\docType{data}
\name{exl_std}
\alias{exl_std}
\title{Standard colors in Excel}
\format{An object of class \code{character} of length 10.}
\usage{
exl_std
}
\description{
Standard colors in Excel
}
\keyword{datasets}
