% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobek-case-info.R
\name{get_file_path}
\alias{get_file_path}
\title{Get file path from sobek case name}
\usage{
get_file_path(case.name = NULL, sobek.project = NULL, type = NULL)
}
\arguments{
\item{case.name}{Name of the case}

\item{sobek.project}{Path to Sobek project folder}

\item{type}{Type of file to get path (lat.dat, bnd.dat, reach, node, structure...)}
}
\value{
Path to the needed file
}
\description{
Get file path from sobek case name
}
\details{
type -> file
\itemize{
\item bnd.dat: get path to boundary.dat
\item lat.dat: get path to lateral.dat
\item reach: get path to reachseg.his
\item node: get path to calcpnt.his
\item structure: get path to struc.his
\item measstation: get path to measstat.his
\item trigger: get path to trigger.def
\item control.def: get path to control.def
\item setting: get path to settings.dat
\item profile.dat: get path to profile.dat
\item profile.def: get path to profile.def
\item struct.dat: get path to struct.dat
\item struct.def: get path to struct.def
}
}
