% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobek-case-info.R
\name{get_file_path}
\alias{get_file_path}
\title{Get file path from sobek case name}
\usage{
get_file_path(case.name = NULL, sobek.project = NULL, type = NULL)
}
\arguments{
\item{case.name}{Name of the case}

\item{sobek.project}{Path to Sobek project folder}

\item{type}{Type of file to get path (lat.dat, bnd.dat, reach, node, structure...)}
}
\value{
Path to the needed file
}
\description{
Get file path from sobek case name
}
\details{
type -> file
\itemize{
\item bnd.dat: get path to boundary.dat
\item lat.dat: get path to lateral.dat
\item reach: get path to reachseg.his
\item node: get path to calcpnt.his
\item structure: get path to struc.his
\item measstation: get path to measstat.his
\item control: get path to controll.def
\item trigger: get path to trigger.def
\item setting: get path to settings.dat
}
}
\examples{
get_file_path(case.name = 'Default', sobek.project = 'd:/so21302/rhein.lit', type = 'bnd.dat')
}
