% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/his-from-case.R
\name{his_from_case}
\alias{his_from_case}
\title{Export data for nodes/reaches for multiple SOBEK Cases}
\usage{
his_from_case(case.list = NULL, sobek.project = NULL, param = 1,
  verbose = TRUE, get.max = FALSE, get.abs.max = FALSE,
  do.par = FALSE, ...)
}
\arguments{
\item{case.list}{Path to the file containing SOBEK Cases to work with}

\item{sobek.project}{Path to Sobek Project folder}

\item{param}{The parameter in the .HIS file (waterlevel, discharge...)}

\item{verbose}{Default TRUE, to display some messages}

\item{get.max}{Default FALSE. If TRUE, the max values will return}

\item{get.abs.max}{Default FALSE. If TRUE, the values that have absolute max will return.
If get.abs.max and get.max both are TRUE, the absolute max will return}

\item{...}{This accept only one parameter in synctax of ID_TYPE = ID_LIST.
ID_TYPE is one of wID, qID, mID, lID (latID), sID, pID, tID
ID_LIST is a character vector.
For example mID = c('p_koeln', 'p_mainz')}

\item{id_name}{Default NULL. To set names for the IDs}
}
\value{
a data.table
}
\description{
For all .HIS files subfolders.
}
