% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/his-from-case.R
\name{his_from_case}
\alias{his_from_case}
\title{Export data for nodes/reaches for multiple SOBEK Cases}
\usage{
his_from_case(
  case.list,
  sobek.project,
  param = 1,
  verbose = TRUE,
  get.max = FALSE,
  get.abs.max = FALSE,
  do.par = FALSE,
  f.comment = "#",
  f.header = FALSE,
  id.names = NULL,
  f.sep = "\\t",
  case.desc = NULL,
  id.type = NULL,
  ...
)
}
\arguments{
\item{case.list}{Path to the file containing SOBEK Cases to work with}

\item{sobek.project}{Path to Sobek Project folder}

\item{param}{The parameter in the .HIS file (waterlevel, discharge...)}

\item{verbose}{Default TRUE, to display some messages}

\item{get.max}{Default FALSE. If TRUE, the max values will return}

\item{get.abs.max}{Default FALSE. If TRUE, the values that have absolute max will return.
If get.abs.max and get.max both are TRUE, the absolute max will return}

\item{f.comment}{Comment character of the files. Default "#"}

\item{f.header}{Logical. Has id.file header?}

\item{id.names}{Character vector for naming the IDs. Default NULL}

\item{f.sep}{Seperator of the id.file, default = 'TAB'}

\item{case.desc}{For renaming case.list}

\item{id.type}{Type of ID, overwrite the ID_TYPE in ...}

\item{...}{This accept only one parameter in syntax of ID_TYPE = ID_LIST.
ID_TYPE is one of wID, qID, mID, lID (latID), sID, pID, tID...
ID_LIST is a character vector.
For example mID = c('p_koeln', 'p_mainz')
\tabular{ll}{
\strong{ID_TYPE} \tab \strong{DESCRIPTION} \cr
mID          \tab Results at Measurements  \cr
wID          \tab Results at Nodes         \cr
qID          \tab Results at Reaches       \cr
lID or latID \tab Results at Laterals      \cr
sID          \tab Results at Structures    \cr
pID          \tab Results for Pumpstations \cr
tID          \tab Results for Triggers     \cr
fmID         \tab Results from Flowmap     \cr
fhID         \tab Results from Flowhis.his \cr
moID         \tab Results from Morpmap.his \cr
smID         \tab Results from Gsedmap.his \cr
shID         \tab Results from Gsedhis.his \cr
File name    \tab HIS file name without '.HIS'. Ex. reachvol = c('ID1', 'ID2')
}}
}
\value{
a data.table
}
\description{
For all .HIS files subfolders.
}
