% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/his-from-case.R
\name{his_from_case}
\alias{his_from_case}
\title{Export data for nodes/reaches for multiple SOBEK Cases}
\usage{
his_from_case(case.list = "", sobek.project = "", param = 1,
  qID = NULL, wID = NULL, lID = NULL, sID = NULL, mID = NULL,
  f.sep = "\\t", f.header = FALSE, f.comment = "#",
  copy.his = FALSE, out.folder = ".", verbose = TRUE)
}
\arguments{
\item{case.list}{Path to the file containing SOBEK Cases to work with}

\item{sobek.project}{Path to Sobek Project folder}

\item{param}{Index of the Paramter to get the data, default = 1}

\item{qID}{Path to file containing list of reach IDs for struct.his}

\item{wID}{Path to file containing list of node IDs for calcpnt.his}

\item{lID}{Path to file containing list of node IDs for lateral.his}

\item{sID}{Path to file containing list of node IDs for struct.his}

\item{mID}{Path to file containing list of node IDs for measstat.his
IDs list shoulde have the first column with IDs(col.name = "sobek_id")
Second column (col.name = "name") is for Names correspond to the IDs}

\item{f.sep}{Seperator of the id.file, default = 'TAB'}

\item{f.header}{Logical. Has id.file header?}

\item{f.comment}{Comment character of the files. Default "#"}

\item{copy.his}{Should .HIS files copied to the output folders?}

\item{out.folder}{Parent folder for the output.}

\item{verbose}{Should message be displayed?}
}
\value{
A list of data.table
}
\description{
For all .HIS files subfolders.
}
