% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-his.R
\name{his_location}
\alias{his_location}
\title{Read location table from .HIS file
Locations are SOBEK's internal index of the node/reach IDs
sobek.ids are IDs of nodes/reaches in the River Network
sobek.id are automatically truncated to max. length of 20 characters by SOBEK}
\usage{
his_location(his.file = "")
}
\arguments{
\item{his.file}{Path to the .HIS file}
}
\value{
a data.table with two column: location & sobek.id
}
\description{
Read location table from .HIS file
Locations are SOBEK's internal index of the node/reach IDs
sobek.ids are IDs of nodes/reaches in the River Network
sobek.id are automatically truncated to max. length of 20 characters by SOBEK
}
