% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_polder.R
\name{plot_polder}
\alias{plot_polder}
\title{Plot hydrographs for a measure with comparing by scenario (zustand)}
\usage{
plot_polder(name = NULL, case.list = NULL, case.desc = case.list,
  sobek.project = NULL, master.tbl = NULL, param = "discharge",
  q.in = FALSE, q.out = FALSE, w.canal = FALSE, ref.mID = NULL,
  y2.scale = 25, h.lines = NULL, peak.nday = NULL,
  peak.pegel = FALSE, lt.by = "zustand", compare.by = NULL,
  facet.by = "hwe", plot.title = NULL, lt.name = "Linienart",
  color.name = "Farbe", text.pos.x = 0, text.pos.y = 1,
  date.break = "3 days", date.label = "\%d.\%m.\%y", text.size = 12,
  text.x.angle = 0L, polder.f = NULL, polder.z = NULL,
  verbose = TRUE)
}
\arguments{
\item{name}{Name of the measure (with/without the measure)}

\item{case.list}{List of cases}

\item{case.desc}{Correct (according to case naming standard in NHWSP) version of case.list, it will be used for legend}

\item{sobek.project}{Path to sobek project}

\item{master.tbl}{Table of ID Coding of the sobek network}

\item{param}{'Waterlevel' or 'Discharge' Hydrograph}

\item{q.in}{Logical. Should discharge through the inlet be plotted?}

\item{q.out}{Logical. Should discharge through the outlet be plotted?}

\item{w.canal}{Logical. Should Wt line inside the measure be plotted?}

\item{ref.mID}{ID of Bezugspegel}

\item{y2.scale}{Scaling between main and secondary y-axes. This is an important paramter. If the line for secondary axis is too big, try to change y2.scale}

\item{h.lines}{List of (name = value) to be displayed as horizontal line}

\item{peak.nday}{Should the plot limit to nday before and after the peak. Default is not (NULL). Otherwise please give a number of days}

\item{peak.pegel}{If the plot should be limit to the peak area, should it be the peak of the referenced location (ref.mID). Default is not.}

\item{lt.by}{Which of (zustand, hwe, vgf...) will be used to show different linetypes?
Others make only sense when your scenarios has same 'zustand' prefix}

\item{facet.by}{Should the graphic be facetted. Default by 'hwe'}

\item{plot.title}{Title of the graphic}

\item{lt.name}{Name of the linetype legend}

\item{color.name}{Name of the color legend}

\item{text.pos.x}{Position of the text block on x-axis, relative value}

\item{text.pos.y}{Position of the text block on y1-axis, absolute or relative value}

\item{date.break}{Breaks for x-axis ('n days', '12 hours'...)}

\item{date.label}{Label format for x-axis. Default dd.mm.yy}

\item{text.size}{Text size for the whole graphic. Default 12}

\item{text.x.angle}{Angle of x-axis text. Default 0}

\item{polder.f}{Area of the measure, for calculating Volume}

\item{polder.z}{Bottom level of the measure for calculating Volume. Give 'auto' for getting the minimum waterlevel in canal value. In this case, make sure the canal is completely dry at T0}

\item{versbose}{Boring? Should some message be displayed?}
}
\value{
A ggplot2 graphic
}
\description{
Plot hydrographs for a measure with comparing by scenario (zustand)
}
