% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_polder}
\alias{plot_polder}
\title{Plot Q/W hydrograph for the}
\usage{
plot_polder(pos = NULL, indt = NULL, peak.duration = 5L,
  x.lab = "Lage", y.lab = "Abfluss (m³/s)", title = "",
  x.angle = 90L, x.hjust = 1L, size = 1L, y.min = NULL,
  delta = FALSE, h.line = NULL, y.interval = 500L, faced = NULL)
}
\arguments{
\item{pos}{Names of columns to draw}

\item{indt}{Input data}

\item{peak.duration}{Halftime of the flooding period, default 5 days}

\item{x.lab}{Label for x axis, default 'Lage'}

\item{y.lab}{Label for y axis, default 'Discharge (m³/s)'}

\item{title}{Graphic title}

\item{x.angle}{Rotation angle of x label}

\item{x.hjust}{hjust of x label}

\item{size}{size of the line}

\item{y.min}{ymin}

\item{delta}{Annotate Delta value, if only for one case and two columns}

\item{h.line}{Numeric vector for adding HLINE}

\item{y.interval}{Tick interval for y axis}

\item{faced}{faced by case, default FALSE}
}
\value{
a ggplot2 graphic
}
\description{
Plot Q/W hydrograph for the
}
