% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_wirkung}
\alias{plot_wirkung}
\title{Quick plot for long profile or pegel (column name)}
\usage{
plot_wirkung(indt = NULL, pegel = "longprofile", sort = TRUE)
}
\arguments{
\item{indt}{Input data.table or data.frame}

\item{sort}{to short longprofile by max value, not yet implemented}
}
\description{
Quick plot for long profile or pegel (column name)
}
