% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobek-simulation.R
\name{sobek_view}
\alias{sobek_view}
\title{Run Sobek Simulation for single case}
\usage{
sobek_view(case.name = NULL, sobek.project = NULL, sobek.path = NULL,
  clear.temp = TRUE, external = TRUE)
}
\arguments{
\item{case.name}{Name of the Case}

\item{sobek.project}{Path to Sobek Project Folder}

\item{sobek.path}{Path to Sobek Program Folder (ex. d:/so21302)}

\item{clear.temp}{Should temp directory to be clear?}

\item{external}{If TRUE, R will use a CMD session to start NETTER,
and the current R session is available to user immediately.}
}
\description{
Run Sobek Simulation for single case
}
